+++ validator/remote/src/main/java/tech/pegasys/teku/validator/remote/RemoteValidatorApiHandler.java
@@ -460,9 +460,7 @@ public class RemoteValidatorApiHandler implements RemoteValidatorApiChannel {
         .map(
             validatorLivenessAtEpoch ->
                 new ValidatorLivenessAtEpoch(
-                    validatorLivenessAtEpoch.index,
-                    validatorLivenessAtEpoch.epoch,
-                    validatorLivenessAtEpoch.isLive))
+                    validatorLivenessAtEpoch.index, validatorLivenessAtEpoch.isLive))
         .toList();
   }